add dependency to calloc and crun.
