+++ internal/util/string.go
@@ -17,6 +17,7 @@
 package util
 
 import (
+	"CraneFrontEnd/generated/protos"
 	"fmt"
 	"math"
 	"regexp"
@@ -471,4 +472,56 @@ func RemoveBracketsWithoutDashOrComma(input string) string {
 		}
 	}
 	return output
+}
+
+func SetTaskDependencies(task *protos.TaskToCtld, depStr string) error {
+	if strings.Contains(depStr, ",") && strings.Contains(depStr, "?") {
+		return fmt.Errorf("cannot use both ',' and '?' in the dependency string")
+	}
+	sep := ","
+	if strings.Contains(depStr, "?") {
+		sep = "?"
+	}
+	depend_all := (sep == ",")
+	if task.Dependencies != nil && depend_all != task.Dependencies.DependAll {
+		return fmt.Errorf("cannot merge dependency with different dependency types(, and ?)")
+	}
+	if task.Dependencies == nil {
+		task.Dependencies = &protos.Dependencies{
+			DependAll: depend_all,
+		}
+	}
+
+	depStr = strings.TrimSpace(depStr)
+	depStrList := strings.Split(depStr, sep)
+	for _, subDepStr := range depStrList {
+		dependencies := strings.Split(subDepStr, ":")
+		if len(dependencies) < 2 {
+			return fmt.Errorf("unrecognized dependency string: %s", subDepStr)
+		}
+		condition := new(protos.DependencyType)
+		switch dependencies[0] {
+		case "after":
+			*condition = protos.DependencyType_AFTER
+		case "afterok":
+			*condition = protos.DependencyType_AFTER_OK
+		case "afternotok":
+			*condition = protos.DependencyType_AFTER_NOT_OK
+		case "afterany":
+			*condition = protos.DependencyType_AFTER_ANY
+		default:
+			return fmt.Errorf("unrecognized dependency type: %s", dependencies[0])
+		}
+		for _, dep := range dependencies[1:] {
+			taskId, err := strconv.ParseUint(dep, 10, 32)
+			if err != nil {
+				return fmt.Errorf("invalid task id: %s", dep)
+			}
+			task.Dependencies.Dependencies = append(task.Dependencies.Dependencies, &protos.DependencyCondition{
+				TaskId: uint32(taskId),
+				Type:   *condition,
+			})
+		}
+	}
+	return nil
 }
